refactor(web): move pages consts to enums folder
