+++ packages/web/pages/_error.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import useTranslation from 'next-translate/useTranslation';
 import { AiOutlineArrowLeft } from 'react-icons/ai';
 import { Link } from '../components/Link';
-import { internal as internalPages } from '../utils/consts/pages';
+import { internal as internalPages } from '../utils/enums/pages.enum';
 
 const Error = ({ statusCode = 400 }) => {
 	const { t } = useTranslation(['error']);
